video_generation.py
# video_generation.py
import os
import sys
import time
import random
import cv2
import numpy as np
import tensorflow as tf
import moviepy.editor as mp
from PIL import Image


class VideoGeneration:
    def __init__(self, video_name, video_quality):
        self.video_name = video_name
        self.video_quality = video_quality
        self.frames = []

    def generate_video(self):
        """
        Generate a high-quality video based on the given video_name and video_quality.
        """
        # Load the video model
        video_model = self.load_video_model()
        # Generate the frames
        self.frames = self.generate_frames(video_model)
        # Create the video from the frames
        self.create_video_from_frames()

    def load_video_model(self):
        """
        Load the pre-trained video generation model.
        """
        # Load the video model (e.g., a TensorFlow or Keras model)
        video_model = None
        return tf.keras.models.load_model("my_video_model.h5")

    def generate_frames(self, video_model):
        """
        Generate the frames for the video using the given video model.
        """
        frames = []
        # TODO: Generate the frames using the video model
        # example code:
        for _ in range(10):
            frame = np.random.randint(0, 255, size=(720, 1280, 3), dtype=np.uint8)
            frames.append(frame)
        return frames

    def create_video_from_frames(self):
        """
        Create a video from the generated frames.
        """
        # TODO: Convert the frames into a video
        # example code:
        fps = 30
        height, width, _ = self.frames[0].shape
        fourcc = cv2.VideoWriter_fourcc(*"mp4v")
        video_writer = cv2.VideoWriter()
        video_writer.open(f"{self.video_name}_{self.video_quality}.mp4", fourcc, fps, (width, height), True)
        for frame in self.frames:
            video_writer.write(frame)
        video_writer.release()
        # Save the video to a file
        self.save_video()

    def save_video(self):
        """
        Save the generated video to a file.
        """
        # Define the output file name and path
        output_file_name = f"{self.video_name}_{self.video_quality}.mp4"
        output_file_path = os.path.join("generated_videos", output_file_name)
        # TODO: Save the video to the output_file_path
        # example code:
        os.rename(f"{self.video_name}_{self.video_quality}.mp4", output_file_path)

    def get_video(self):
        """
        Return the generated video.
        """
        # TODO: Return the generated video as a video object or a file path
        # example code:
        return mp.VideoFileClip(f"generated_videos/{self.video_name}_{self.video_quality}.mp4")


if __name__ == "__main__":
    # Example usage:
    video_gen = VideoGeneration("example_video", "high")
    video_gen.generate_video()
    generated_video = video_gen.get_video()
