requirements.txt
requests
Pillow
tensorflow==2.3.1
keras==2.3.1
scikit-learn
numpy
pandas
matplotlib
seaborn
nltk
spacy
gensim
pyaudio
SpeechRecognition
googletrans==4.0.0-rc1
pyttsx3
twilio
sendgrid
Flask
Django
PyQt5
wxPython
cryptography
setuptools
opencv-python